ONCO-64: Moved Exception classes to specific file
